+++ pkg/identityserver/emails/validate.go
@@ -38,16 +38,18 @@ const validateText = `Please confirm your email address for {{.Network.Name}}.
 
 Your email address will be used as contact for {{.Entity.Type}} "{{.Entity.ID}}".
 
-Confirm via web interface:
-{{ .Network.IdentityServerURL }}/validate?reference={{ .ID }}&token={{ .Token }}
+You can go to {{ .Network.IdentityServerURL }}/validate?reference={{ .ID }}&token={{ .Token }} to confirm your email address.
 
-Confirm via command-line interface:
+If you prefer to use the command-line interface, you can run the following command:
+
 ttn-lw-cli {{.Entity.Type}}s contact-info validate {{.ID}} {{.Token}}
+
+For more information on how to use the command-line interface, please refer to the documentation: {{ documentation_url "/getting-started/cli/" }}.
 
 {{- if .TTL }}
 
-These confirmation links expire {{ .FormatTTL }}.
-{{ end -}}
+The confirmation token expires {{ .FormatTTL }}, so confirm your email address before then.
+{{- end }}
 `
 
 // DefaultTemplates returns the default templates for this email.